requirements-dev.txt
bleach>=3.1.1
boto3==1.10.26
botocore==1.13.26
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.8
jeepney==0.4.1
Jinja2==2.10.3
jmespath==0.9.4
keyring==20.0.0
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow>=6.2.2
piexif>=1.1.3
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.5.2
python-dateutil==2.8.0
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
SecretStorage==3.1.1
six==1.13.0
tqdm==4.41.0
twine==3.1.1
urllib3==1.25.7
webencodings==0.5.1
